Make int / int return float
I miiiight have thought implementing this as a integer division (i.e.
3/2 returns 1) made sense at some point, but if we assess this behavior
now, I think it's pretty clearly the wrong decision and would surprise
users, especially given the similatiries we share with Python.

This commit simply converts both ints into floats prior to the division,
resulting in a float answer as well.

+++ core/interpreter_long.go
@@ -34,7 +34,7 @@ func (i *MainInterpreter) executeOp(left interface{}, right interface{}, tkn Tok
 			case OP_MULTIPLY:
 				return coercedLeft * coercedRight
 			case OP_DIVIDE:
-				return coercedLeft / coercedRight
+				return float64(coercedLeft) / float64(coercedRight)
 			case OP_GREATER:
 				return coercedLeft > coercedRight
 			case OP_GREATER_EQUAL:
